moved editnamefield to its own component, standardized parameter order for column actions
